feature-db: put local connector and mysql connector into separate folder to avoid unused import

+++ db/mysql/mysqlconnector.go
@@ -6,6 +6,7 @@ import (
 	"strings"
 	"sync"
 
+	. "github.com/smiecj/go_common/db"
 	"github.com/smiecj/go_common/errorcode"
 	"github.com/smiecj/go_common/util/log"
 	"gorm.io/driver/mysql"
@@ -39,39 +40,43 @@ type mysqlConnector struct {
 
 // mysql: 插入数据
 // 后续: 对批量插入场景，单次插入的数据量进行控制
-func (connector *mysqlConnector) Insert(funcArr ...rdbInsertConfigFunc) (ret updateRet, err error) {
-	action := makeRDBInsertAction()
+func (connector *mysqlConnector) Insert(funcArr ...RDBInsertConfigFunc) (ret UpdateRet, err error) {
+	action := MakeRDBInsertAction()
 	for _, currentFunc := range funcArr {
 		currentFunc(action)
 	}
 
 	// 判断是通过 object 插入 还是通过 直接指定key-value插入
 	var dbRet *gorm.DB
-	if len(action.fieldArr) != 0 {
+	fieldArr := action.GetFieldArr()
+	objectArr := action.GetObjectArr()
+	if len(fieldArr) != 0 {
 		keyValueMapArr := make([]map[string]interface{}, 0)
-		for _, currentField := range action.fieldArr {
+		for _, currentField := range fieldArr {
 			currentKeyValueMap := make(map[string]interface{}, 0)
-			for key, value := range currentField.keyValueMap {
+			for key, value := range currentField.GetMap() {
 				currentKeyValueMap[key] = value
 			}
 			keyValueMapArr = append(keyValueMapArr, currentKeyValueMap)
 		}
-		dbRet = connector.db.Table(action.getSpaceName()).Create(keyValueMapArr)
-	} else if len(action.objectArr) != 0 {
+		dbRet = connector.db.Table(action.GetSpaceName()).Create(keyValueMapArr)
+	} else if len(objectArr) != 0 {
 		searchKeyArr := []string{}
-		if len(action.keyArr) != 0 {
-			searchKeyArr = action.keyArr
+		keyArr := action.GetKeyArr()
+		if len(keyArr) != 0 {
+			searchKeyArr = keyArr
 		}
-		// 注意数组类型需要转换一下，传入的 interface{} 数组无法被 gorm 识别
-		var toInsertArr interface{} = action.objectArr
-		if nil != action.objectArrType {
-			slice := reflect.MakeSlice(action.objectArrType, 0, 0)
-			for _, currentObj := range action.objectArr {
+		// 注意数组类型需要转换一下，传入的 interface{} 数组无法被 gorm 识别（即数组需要保持原有的type）
+		var toInsertArr interface{} = objectArr
+		objectArrType := action.GetObjectArrType()
+		if nil != objectArrType {
+			slice := reflect.MakeSlice(objectArrType, 0, 0)
+			for _, currentObj := range objectArr {
 				slice = reflect.Append(slice, reflect.ValueOf(currentObj))
 			}
 			toInsertArr = slice.Interface()
 		}
-		dbRet = connector.db.Table(action.getSpaceName()).Select(searchKeyArr).Create(toInsertArr)
+		dbRet = connector.db.Table(action.GetSpaceName()).Select(searchKeyArr).Create(toInsertArr)
 	} else {
 		return ret, errorcode.BuildErrorWithMsg(errorcode.DBParamInvalid, "Insert failed: to insert data is empty")
 	}
@@ -79,35 +84,39 @@ func (connector *mysqlConnector) Insert(funcArr ...rdbInsertConfigFunc) (ret upd
 	ret.AffectedRows, err = int(dbRet.RowsAffected), dbRet.Error
 
 	if nil != err {
-		log.Error("[mysqlConnector.Insert] Insert failed: table: %s, reason: %s", action.getSpaceName(), err.Error())
+		log.Error("[mysqlConnector.Insert] Insert failed: table: %s, reason: %s", action.GetSpaceName(), err.Error())
 	} else {
-		log.Info("[mysqlConnector.Insert] Insert success: %s, insert rows: %d", action.getSpaceName(), ret.AffectedRows)
+		log.Info("[mysqlConnector.Insert] Insert success: %s, insert rows: %d", action.GetSpaceName(), ret.AffectedRows)
 	}
 	return
 }
 
 // mysql: 更新数据
-func (connector *mysqlConnector) Update(funcArr ...rdbUpdateConfigFunc) (ret updateRet, err error) {
-	action := makeRDBUpdateAction()
+func (connector *mysqlConnector) Update(funcArr ...RDBUpdateConfigFunc) (ret UpdateRet, err error) {
+	action := MakeRDBUpdateAction()
 	for _, currentFunc := range funcArr {
 		currentFunc(action)
 	}
 
 	// 根据查询条件 更新指定数据，只更新一种取值
 	var dbRet *gorm.DB
-	if len(action.fieldArr) != 0 {
+	fieldArr := action.GetFieldArr()
+	objectArr := action.GetObjectArr()
+	keyArr := action.GetKeyArr()
+	condition := action.GetCondition()
+	if len(fieldArr) != 0 {
 		keyValueMap := make(map[string]interface{}, 0)
-		currentField := action.fieldArr[0]
-		for key, value := range currentField.keyValueMap {
+		currentField := fieldArr[0]
+		for key, value := range currentField.GetMap() {
 			keyValueMap[key] = value
 		}
-		dbRet = connector.db.Table(action.getSpaceName()).Where(action.condition.WhereArr.toSQL()).Updates(keyValueMap)
-	} else if len(action.objectArr) != 0 {
+		dbRet = connector.db.Table(action.GetSpaceName()).Where(condition.WhereArr.ToSQL()).Updates(keyValueMap)
+	} else if len(objectArr) != 0 {
 		searchKeyArr := []string{}
-		if len(action.keyArr) != 0 {
-			searchKeyArr = action.keyArr
+		if len(keyArr) != 0 {
+			searchKeyArr = keyArr
 		}
-		dbRet = connector.db.Table(action.getSpaceName()).Where(action.condition.WhereArr.toSQL()).Select(searchKeyArr).Updates(action.objectArr[0])
+		dbRet = connector.db.Table(action.GetSpaceName()).Where(condition.WhereArr.ToSQL()).Select(searchKeyArr).Updates(objectArr[0])
 	} else {
 		return ret, errorcode.BuildErrorWithMsg(errorcode.DBParamInvalid, "Insert failed: to insert data is empty")
 	}
@@ -115,49 +124,54 @@ func (connector *mysqlConnector) Update(funcArr ...rdbUpdateConfigFunc) (ret upd
 	ret.AffectedRows, err = int(dbRet.RowsAffected), dbRet.Error
 
 	if nil != err {
-		log.Error("[mysqlConnector.Update] Update failed, table: %s, reason: %s", action.getSpaceName(), err.Error())
+		log.Error("[mysqlConnector.Update] Update failed, table: %s, reason: %s", action.GetSpaceName(), err.Error())
 	} else {
-		log.Info("[mysqlConnector.Update] Update success: %s, update rows: %d", action.getSpaceName(), ret.AffectedRows)
+		log.Info("[mysqlConnector.Update] Update success: %s, update rows: %d", action.GetSpaceName(), ret.AffectedRows)
 	}
 	return
 }
 
 // mysql: 删除数据
-func (connector *mysqlConnector) Delete(funcArr ...rdbDeleteConfigFunc) (ret updateRet, err error) {
-	action := makeRDBDeleteAction()
+func (connector *mysqlConnector) Delete(funcArr ...RDBDeleteConfigFunc) (ret UpdateRet, err error) {
+	action := MakeRDBDeleteAction()
 	for _, currentFunc := range funcArr {
 		currentFunc(action)
 	}
+
+	condition := action.GetCondition()
 	limitCondition := ""
-	if action.condition.Limit != 0 {
-		limitCondition = fmt.Sprintf("limit %d", action.condition.Limit)
+	if condition.Limit != 0 {
+		limitCondition = fmt.Sprintf("limit %d", condition.Limit)
 	}
 
 	dbRet := connector.db.Exec(fmt.Sprintf("DELETE FROM %s WHERE %s %s",
-		action.getSpaceName(), action.condition.WhereArr.toSQL(), limitCondition))
+		action.GetSpaceName(), condition.WhereArr.ToSQL(), limitCondition))
 	ret.AffectedRows, err = int(dbRet.RowsAffected), dbRet.Error
 
 	if nil != err {
-		log.Error("[mysqlConnector.Delete] Delete failed, table: %s, reason: %s", action.getSpaceName(), err.Error())
+		log.Error("[mysqlConnector.Delete] Delete failed, table: %s, reason: %s", action.GetSpaceName(), err.Error())
 	} else {
-		log.Info("[mysqlConnector.Delete] Delete success: %s, update rows: %d", action.getSpaceName(), ret.AffectedRows)
+		log.Info("[mysqlConnector.Delete] Delete success: %s, update rows: %d", action.GetSpaceName(), ret.AffectedRows)
 	}
 	return
 }
 
 // mysql: 查询数据
-func (connector *mysqlConnector) Search(funcArr ...rdbSearchConfigFunc) (ret searchRet, err error) {
-	action := makeRDBSearchAction()
+func (connector *mysqlConnector) Search(funcArr ...RDBSearchConfigFunc) (ret SearchRet, err error) {
+	action := MakeRDBSearchAction()
 	for _, currentFunc := range funcArr {
 		currentFunc(action)
 	}
 
+	condition := action.GetCondition()
+
 	// 统计 count
 	var count int64
-	dbRet := connector.db.Table(action.getSpaceName()).
-		Select(action.keyArr).Where(action.condition.WhereArr.toSQL()).Count(&count)
+	keyArr := action.GetKeyArr()
+	dbRet := connector.db.Table(action.GetSpaceName()).
+		Select(keyArr).Where(condition.WhereArr.ToSQL()).Count(&count)
 	if nil != dbRet.Error {
-		log.Error("[mysqlConnector.Count] Count failed, table: %s, reason: %s", action.getSpaceName(), err.Error())
+		log.Error("[mysqlConnector.Count] Count failed, table: %s, reason: %s", action.GetSpaceName(), err.Error())
 		return ret, dbRet.Error
 	} else {
 		ret.Total = int(count)
@@ -165,23 +179,24 @@ func (connector *mysqlConnector) Search(funcArr ...rdbSearchConfigFunc) (ret sea
 
 	// order condition
 	var orderStr string
-	if "" != action.condition.Order.Field {
-		orderStr = fmt.Sprintf("%s %s", action.condition.Order.Field, action.condition.Order.Sc)
+	if "" != condition.Order.Field {
+		orderStr = fmt.Sprintf("%s %s", condition.Order.Field, condition.Order.Sc)
 	}
 
-	if nil != action.objectArrType {
-		objectReflectArr := reflect.MakeSlice(action.objectArrType, 0, 0).Interface()
-		dbRet = connector.db.Table(action.getSpaceName()).
-			Select(action.keyArr).Where(action.condition.WhereArr.toSQL()).Order(orderStr).
-			Offset(action.condition.Page.No * action.condition.Page.Limit).Limit(action.condition.Page.Limit).
+	objectArrType := action.GetObjectArrType()
+	if nil != objectArrType {
+		objectReflectArr := reflect.MakeSlice(objectArrType, 0, 0).Interface()
+		dbRet = connector.db.Table(action.GetSpaceName()).
+			Select(keyArr).Where(condition.WhereArr.ToSQL()).Order(orderStr).
+			Offset(condition.Page.No * condition.Page.Limit).Limit(condition.Page.Limit).
 			Find(&objectReflectArr)
 		ret.ObjectArr = objectReflectArr
 	} else {
 		// 非导入到 object 情况，存在 value 在转换的时候不准确的问题，需要测试
 		keyValueMapArr := make([]map[string]interface{}, 0)
-		dbRet = connector.db.Table(action.getSpaceName()).
-			Select(action.keyArr).Where(action.condition.WhereArr.toSQL()).Order(orderStr).
-			Offset(action.condition.Page.No * action.condition.Page.Limit).Limit(action.condition.Page.Limit).
+		dbRet = connector.db.Table(action.GetSpaceName()).
+			Select(keyArr).Where(condition.WhereArr.ToSQL()).Order(orderStr).
+			Offset(condition.Page.No * condition.Page.Limit).Limit(condition.Page.Limit).
 			Find(&keyValueMapArr)
 		for _, currentKeyValueMap := range keyValueMapArr {
 			currentField := BuildNewField()
@@ -189,72 +204,74 @@ func (connector *mysqlConnector) Search(funcArr ...rdbSearchConfigFunc) (ret sea
 				valuestr := fmt.Sprintf("%v", value)
 				currentField.AddKeyValue(key, valuestr)
 			}
-			ret.addField(currentField)
+			ret.AddField(currentField)
 		}
 	}
 
 	ret.Len, err = int(dbRet.RowsAffected), dbRet.Error
 
 	if nil != err {
-		log.Error("[mysqlConnector.Select] Select failed, table: %s, reason: %s", action.getSpaceName(), err.Error())
+		log.Error("[mysqlConnector.Select] Select failed, table: %s, reason: %s", action.GetSpaceName(), err.Error())
 	} else {
-		log.Info("[mysqlConnector.Select] Select success: %s, search rows: %d", action.getSpaceName(), ret.Len)
+		log.Info("[mysqlConnector.Select] Select success: %s, search rows: %d", action.GetSpaceName(), ret.Len)
 	}
 	return
 }
 
 // mysql: 统计数据量
-func (connector *mysqlConnector) Count(funcArr ...rdbSearchConfigFunc) (ret searchRet, err error) {
-	action := makeRDBSearchAction()
+func (connector *mysqlConnector) Count(funcArr ...RDBSearchConfigFunc) (ret SearchRet, err error) {
+	action := MakeRDBSearchAction()
 	for _, currentFunc := range funcArr {
 		currentFunc(action)
 	}
 
+	condition := action.GetCondition()
 	var count int64
-	dbRet := connector.db.Table(action.getSpaceName()).Where(action.condition.WhereArr.toSQL()).Count(&count)
+	dbRet := connector.db.Table(action.GetSpaceName()).Where(condition.WhereArr.ToSQL()).Count(&count)
 
 	ret.Total, err = int(count), dbRet.Error
 
 	if nil != err {
-		log.Error("[mysqlConnector.Count] Count failed, table: %s, reason: %s", action.getSpaceName(), err.Error())
+		log.Error("[mysqlConnector.Count] Count failed, table: %s, reason: %s", action.GetSpaceName(), err.Error())
 	} else {
-		log.Info("[mysqlConnector.Count] Count success: %s, total: %d", action.getSpaceName(), ret.Total)
+		log.Info("[mysqlConnector.Count] Count success: %s, total: %d", action.GetSpaceName(), ret.Total)
 	}
 	return
 }
 
 // mysql: distinct
-func (connector *mysqlConnector) Distinct(funcArr ...rdbSearchConfigFunc) (ret searchRet, err error) {
-	action := makeRDBSearchAction()
+func (connector *mysqlConnector) Distinct(funcArr ...RDBSearchConfigFunc) (ret SearchRet, err error) {
+	action := MakeRDBSearchAction()
 	for _, currentFunc := range funcArr {
 		currentFunc(action)
 	}
 	// distinct 必须指定需要查询的列名
-	if len(action.keyArr) == 0 {
+	keyArr := action.GetKeyArr()
+	if len(keyArr) == 0 {
 		return ret, errorcode.BuildErrorWithMsg(errorcode.DBParamInvalid, "[mysqlConnector.Distinct] distinct must set key array")
 	}
 
 	fieldValueArr := make([]string, 0)
 	// 查询包含多个字段，通过 SQL concat 关键字进行拼接
 	var distinctColumn string
-	if 1 == len(action.keyArr) {
-		distinctColumn = action.keyArr[0]
+	if 1 == len(keyArr) {
+		distinctColumn = keyArr[0]
 	} else {
 		distinctColumn = "CONCAT("
-		for index := 0; index < len(action.keyArr); index++ {
+		for index := 0; index < len(keyArr); index++ {
 			if index != 0 {
 				distinctColumn += fmt.Sprintf(", '%s', ", distinctSeparator)
 			}
-			distinctColumn += fmt.Sprintf("%s", action.keyArr[index])
+			distinctColumn += fmt.Sprintf("%s", keyArr[index])
 		}
 		distinctColumn += ")"
 	}
 
-	dbRet := connector.db.Table(action.getSpaceName()).Select(action.keyArr).Where(action.condition.WhereArr.toSQL()).
+	dbRet := connector.db.Table(action.GetSpaceName()).Select(keyArr).Where(action.GetCondition().WhereArr.ToSQL()).
 		Distinct().Pluck(distinctColumn, &fieldValueArr)
 	err = dbRet.Error
 	if nil != dbRet.Error {
-		log.Error("[mysqlConnector.Distinct] Distinct %s get field value failed: %s", action.getSpaceName(), err.Error())
+		log.Error("[mysqlConnector.Distinct] Distinct %s get field value failed: %s", action.GetSpaceName(), err.Error())
 		return ret, err
 	}
 
@@ -262,10 +279,10 @@ func (connector *mysqlConnector) Distinct(funcArr ...rdbSearchConfigFunc) (ret s
 	for _, currentValue := range fieldValueArr {
 		currentField := BuildNewField()
 		valueSplitArr := strings.Split(currentValue, distinctSeparator)
-		for index := 0; index < len(action.keyArr); index++ {
-			currentField.AddKeyValue(action.keyArr[index], valueSplitArr[index])
+		for index := 0; index < len(keyArr); index++ {
+			currentField.AddKeyValue(keyArr[index], valueSplitArr[index])
 		}
-		ret.addField(currentField)
+		ret.AddField(currentField)
 	}
 	// distinct 只计算 len，不计算 total
 	ret.Len = len(fieldValueArr)
@@ -276,7 +293,7 @@ func (connector *mysqlConnector) Distinct(funcArr ...rdbSearchConfigFunc) (ret s
 func GetMySQLConnector(option MySQLConnectOption) RDBConnector {
 	var connector RDBConnector
 	mysqlConnectorLock.RLock()
-	if nil == fileConnectorMap {
+	if nil == mysqlConnectorMap {
 		mysqlConnectorMap = make(map[MySQLConnectOption]RDBConnector)
 	}
 	connector = mysqlConnectorMap[option]
@@ -286,8 +303,8 @@ func GetMySQLConnector(option MySQLConnectOption) RDBConnector {
 		return connector
 	}
 
-	fileConnectorLock.Lock()
-	defer fileConnectorLock.Unlock()
+	mysqlConnectorLock.Lock()
+	defer mysqlConnectorLock.Unlock()
 
 	// useAffectedRows 等配置提示无效，后续需要确认原因
 	extendParam := ""
